src/requirements.txt
matplotlib==3.10.8
numpy==2.4.2
Pillow==12.1.1
processing==0.52

# Generated with pipreqs